+++ app/controllers/federails/server/nodeinfo_controller.rb
@@ -12,7 +12,7 @@ module Federails
 
         @total = @active_halfyear = @active_month = 0
         @has_user_counts = false
-        Federails::Configuration.entity_types.each_value do |config|
+        Federails::Configuration.actor_types.each_value do |config|
           next unless (method = config[:user_count_method]&.to_sym)
 
           @has_user_counts = true