+++ app/models/concerns/federails/actor_entity.rb
@@ -27,9 +27,9 @@ module Federails
       has_one :federails_actor, class_name: 'Federails::Actor', as: :entity, dependent: :destroy
 
       after_create :create_federails_actor, if: lambda {
-        raise("Entity not configured for #{self.class.name}. Did you use \"acts_as_federails_actor\"?") unless Federails::Configuration.entity_types.key? self.class.name
+        raise("Entity not configured for #{self.class.name}. Did you use \"acts_as_federails_actor\"?") unless Federails::Configuration.actor_types.key? self.class.name
 
-        Federails::Configuration.entity_types[self.class.name][:auto_create_actors]
+        Federails::Configuration.actor_types[self.class.name][:auto_create_actors]
       }
 
       # Configures the mapping between entity and actor