+++ CHANGELOG.md
@@ -37,6 +37,7 @@ As we're still in a kind of early development, some changes to Federails interna
 - Methods included in `Federails::Entity` are renamed with `federails` in them to avoid confusion and make projects with _actors_ able to use the gem
 - [**BREAKING**] Concern `Federails::Entity` has been renamed to `Federails::ActorEntity`
 - Internal method `Federails::Configuration.register_entity` has been renamed to `Federails::Configuration.register_actor_class`.
+- [**BREAKING**] Configuration key `Federails::Configuration.entity_types` has been renamed to `Federails::Configuration.actor_types`
 
 ## [0.3.0] 2024-11-23